requirements.txt
# Google ADK
google-adk
google-genai
google-generativeai

# Ambiente e configuração
python-dotenv

# Interface web
flask
